chore: remove init_template tool, run configs; update turbo_widgets layout
- Delete turbo_template init_template Dart tool (replacers, utils, tests)
- Delete .run/ and turbo_template/.run/ IDE run configurations
- Update t_bento_grid and t_calculated_height in turbo_widgets

Made-with: Cursor

+++ turbo_widgets/lib/src/widgets/layout/t_bento_grid.dart
@@ -190,7 +190,7 @@ class TBentoGrid extends StatefulWidget {
   final double? maxWidth;
 
   final double baseHeight;
-  final int multiplierThreshold;
+  final double multiplierThreshold;
   final double calculatedMinHeight;
   final double? calculatedMaxHeight;
